Move Interface::Block implementation
