+++ lib/steep.rb
@@ -43,6 +43,7 @@ require "steep/ast/builtin"
 require "steep/ast/types/factory"
 
 require "steep/interface/function"
+require "steep/interface/block"
 require "steep/interface/method_type"
 require "steep/interface/substitution"
 require "steep/interface/interface"
